# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for NexFlux Virtual Lab - An interactive electronics learning platform",
        "title": "NexFlux Virtual Lab API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "idnmakerspace@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/achievements": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get list of all achievements with user's unlock status and progress",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gamification"
                ],
                "summary": "Get all achievements",
                "responses": {
                    "200": {
                        "description": "List of achievements with stats",
                        "schema": {
                            "$ref": "#/definitions/dto.AchievementListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/achievements/user": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get list of achievements unlocked by the current user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gamification"
                ],
                "summary": "Get my achievements",
                "responses": {
                    "200": {
                        "description": "List of unlocked achievements",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/leaderboard": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get leaderboard rankings (weekly, monthly, or all-time)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gamification"
                ],
                "summary": "Get leaderboard",
                "parameters": [
                    {
                        "type": "string",
                        "default": "weekly",
                        "description": "Leaderboard type (weekly, monthly, all_time)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of entries to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Leaderboard with rankings",
                        "schema": {
                            "$ref": "#/definitions/dto.LeaderboardResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/streak": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get user's current streak information and history",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gamification"
                ],
                "summary": "Get streak info",
                "responses": {
                    "200": {
                        "description": "Streak information",
                        "schema": {
                            "$ref": "#/definitions/dto.StreakResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/users/me/stats": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get comprehensive user gamification statistics including XP, level, streak, etc.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gamification"
                ],
                "summary": "Get user stats",
                "responses": {
                    "200": {
                        "description": "User statistics",
                        "schema": {
                            "$ref": "#/definitions/dto.UserStatsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AchievementListResponse": {
            "type": "object",
            "properties": {
                "achievements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AchievementResponse"
                    }
                },
                "stats": {
                    "type": "object",
                    "properties": {
                        "total": {
                            "type": "integer"
                        },
                        "unlocked": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "dto.AchievementResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_unlocked": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "progress": {
                    "type": "integer"
                },
                "rarity": {
                    "type": "string"
                },
                "requirement_type": {
                    "type": "string"
                },
                "requirement_value": {
                    "type": "integer"
                },
                "unlocked_at": {
                    "type": "string"
                },
                "xp_reward": {
                    "type": "integer"
                }
            }
        },
        "dto.AchievementStats": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "unlocked": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ChallengeStats": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "integer"
                },
                "in_progress": {
                    "type": "integer"
                },
                "total_available": {
                    "type": "integer"
                }
            }
        },
...
        "dto.LeaderboardEntryResponse": {
            "type": "object",
            "properties": {
                "is_current_user": {
                    "type": "boolean"
                },
                "rank": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/dto.UserBriefResponse"
                },
                "xp": {
                    "type": "integer"
                }
            }
        },
        "dto.LeaderboardResponse": {
            "type": "object",
            "properties": {
                "current_user_rank": {
                    "type": "integer"
                },
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.LeaderboardEntryResponse"
                    }
                },
                "period": {
                    "$ref": "#/definitions/dto.PeriodResponse"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "dto.PeriodResponse": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                }
            }
        },
...
        "dto.ProjectStats": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "integer"
                },
                "in_progress": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.StreakInfo": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "last_active": {
                    "type": "string"
                },
                "longest": {
                    "type": "integer"
                }
            }
        },
        "dto.StreakResponse": {
            "type": "object",
            "properties": {
                "current_streak": {
                    "type": "integer"
                },
                "is_active_today": {
                    "type": "boolean"
                },
                "last_activity_date": {
                    "type": "string"
                },
                "longest_streak": {
                    "type": "integer"
                },
                "streak_protects": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UserBriefResponse": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "dto.UserStatsResponse": {
            "type": "object",
            "properties": {
                "achievements": {
                    "$ref": "#/definitions/dto.AchievementStats"
                },
                "challenges": {
                    "$ref": "#/definitions/dto.ChallengeStats"
                },
                "current_xp": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "projects": {
                    "$ref": "#/definitions/dto.ProjectStats"
                },
                "streak": {
                    "$ref": "#/definitions/dto.StreakInfo"
                },
                "target_xp": {
                    "type": "integer"
                },
                "total_xp": {
                    "type": "integer"
                },
                "xp_this_week": {
                    "type": "integer"
                },
                "xp_to_next_level": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}